chapter 5 L5

+++ src/copystatic.py
@@ -2,7 +2,8 @@ import os
 import shutil
     
 def copystatic(dir_path, destination_path):
-    shutil.rmtree(destination_path)
+    if os.path.isdir(destination_path):
+        shutil.rmtree(destination_path)
     os.mkdir(destination_path)
     copy_rec(dir_path, destination_path)
